# api/openapi.yaml
openapi: 3.1.0
info:
  title: Jasper API
  version: 0.2.0
  description: |
    Jasper HTTP API: notes CRUD keyed by ULID, the file-tree projection,
    admin endpoints (migration status, full re-index) backing DATA-10, and
    tags, backlinks and wiki-link search-titles
    (TAGS-03, TAGS-04, TAGS-06, TAGS-07, LINKS-06, LINKS-08).

    Source of truth: every route Jasper exposes lives in this file (API-01).
    Adding a route here without a Go handler causes a compile error
    under oapi-codegen StrictServerInterface (API-02).
servers:
  - url: /api/v1
    description: Same-origin (Go binary or Vite dev proxy)
...
paths:
...
  /config:
    get:
      tags: [config]
      operationId: getConfig
      summary: Read user configuration
      description: |
        Returns the current config.json contents (DESIGN.md §11). Fields the
        UI does not surface still ship with sensible defaults.
      responses:
        "200":
          description: Current config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Config" }
    put:
      tags: [config]
      operationId: putConfig
      summary: Replace user configuration (whole-document)
      description: |
        Writes config.json atomically via fsstore.AtomicWrite
        (DATA-13: temp + fsync + rename + fsync(parent)).
        Strict decode: unknown fields → 400 invalid_request.
        Replaces the whole document — retained for whole-document writes
        (per-section Reset, first-run fixtures). For partial writes that
        should leave every other field untouched, use PATCH /config.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Config" }
      responses:
        "200":
          description: Config saved; echoes the persisted document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Config" }
        "400":
          description: Invalid request (malformed body, unknown field, out-of-range numeric)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Save failed (filesystem error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      tags: [config]
      operationId: patchConfig
      summary: Partially update user configuration
      description: |
        Writes only the keys present in the request body via
        config.SaveMergedPartial; fields absent from the body are left
        untouched on disk (including unmanaged/hand-added keys). Writes
        are serialised server-side so concurrent PATCH/PUT requests never
        lose an update. Strict decode: unknown fields → 400 invalid_request,
        same as PUT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigPatch" }
      responses:
        "200":
          description: Config saved; echoes the full persisted document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Config" }
        "400":
          description: Invalid request (malformed body, unknown field, out-of-range numeric)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Save failed (filesystem error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
...
  schemas:
...
    Config:
      type: object
      required: [appName, dailyNotes, editor, theme]
      additionalProperties: false   # strict decode — unknown keys → 400
      # The `server` and `mcp` blocks are NOT in `required` so older
      # config.json files on disk continue to load; the Go side fills
      # defaults in load.go when the blocks are absent.
      # `additionalProperties: false` still holds — PUT /config payloads
      # carrying `server` or `mcp` round-trip cleanly. PUT bodies missing
      # them are accepted (strict validator pre-fills defaults).
      properties:
        appName:
          type: string
          minLength: 1
          maxLength: 64
          default: "Jasper"
        theme:
          type: string
          enum: [dark, light]
          default: dark
        dailyNotes:
          type: object
          required: [template]
          additionalProperties: false
          properties:
            template:
              type: string
              maxLength: 1024
              default: ""
        editor:
          type: object
          required: [fontSize, lineHeight, autosaveMs]
          additionalProperties: false
          properties:
            fontSize:
              type: integer
              minimum: 8
              maximum: 32
              default: 15
            lineHeight:
              type: number
              format: double
              minimum: 1
              maximum: 3
              default: 1.6
            autosaveMs:
              type: integer
              minimum: 250
              maximum: 10000
              default: 2000
              description: |
                Autosave debounce interval in milliseconds. Replaces the
                hard-coded 2000ms constant in EditorPane.tsx.
                Read at mount; requires reload to apply.
            showProperties:
              # Toggles visibility of the properties table above the note
              # body. Optional; defaults true.
              type: boolean
              default: true
            autoPair:
              # EDIT-01 — auto-close brackets/quotes in CM6.
              type: boolean
              default: true
            foldGutter:
              # EDIT-03 — heading/list code-folding gutter.
              type: boolean
              default: true
            lineNumbers:
              # EDIT-03 — CM6 line-number gutter.
              type: boolean
              default: false
            lineWidth:
              # EDIT-04 — maximum width of the writing column, in CSS pixels.
              type: integer
              minimum: 400
              maximum: 2000
              default: 700
        accent:
          type: string
          enum: [purple, sky, green, orange]
          default: purple
          description: |
            UI accent color. Applied to interactive elements (links, buttons,
            active tab indicator). Optional — legacy configs without this key
            load with the default "purple". Coerced to "purple" on load if
            the value is not a member of the enum.
        readingFont:
          type: string
          enum: [sans, serif]
          default: sans
          description: |
            Font family used in the reading/preview surface. "sans" uses the
            system sans-serif stack; "serif" uses the system serif stack.
            Optional — legacy configs without this key load with default "sans".
            Coerced to "sans" on load if the value is not "sans" or "serif".
        server:
          # HTTP listener port + resolved data directory. Port pinned at
          # 6683 (T9 spelling of "NOTE"); the wizard does NOT expose a port
          # override (config-file edit only). DataDir is set by the
          # first-run wizard and re-read at every server boot.
          # NET-01: bind defaults to loopback; 0.0.0.0 exposes all
          # interfaces. MCP always stays loopback regardless.
          type: object
          required: [port, dataDir, bind]
          additionalProperties: false
          properties:
            port:
              type: integer
              minimum: 1
              maximum: 65535
              default: 6683
            dataDir:
              type: string
              minLength: 1
            bind:
              type: string
              default: "127.0.0.1"
              description: |
                HTTP listener bind address. Defaults to "127.0.0.1" (loopback-only).
                Set to "0.0.0.0" to expose on all network interfaces. Requires a
                server restart. MCP always remains loopback regardless of this value.
        mcp:
          # The MCP listener always starts on boot; there is no
          # enable/disable toggle. AI write access is governed solely by
          # per-folder write grants. Port pinned at 6684. Bind is
          # loopback-only (127.0.0.1), enforced at listener startup via
          # netbind.RequireLoopbackBind.
          type: object
          required: [port, bind]
          additionalProperties: false
          properties:
            port:
              type: integer
              minimum: 1
              maximum: 65535
              default: 6684
            bind:
              type: string
              default: "127.0.0.1"
            auditLog:
              # MCP2-02 — tool-call audit-log toggle.
              type: boolean
              default: false
        templates:
          # TPL-01 — vault-relative folder holding templates.
          type: object
          required: [folder]
          additionalProperties: false
          properties:
            folder:
              type: string
              maxLength: 64
              default: "Templates"
    ConfigPatch:
      type: object
      additionalProperties: false   # strict decode — unknown keys → 400
      description: |
        Sparse partial-update body for PATCH /config. Every property is
        optional at every nesting level (no `required` anywhere in this
        schema) — that is the entire mechanism for distinguishing "key
        omitted, leave untouched" from "key present with value 0/false/
        empty-string, write it". Only the keys present in the request body
        are written; every other key already on disk (including
        unmanaged/hand-added keys) is preserved.
      properties:
        appName:
          type: string
          minLength: 1
          maxLength: 64
        theme:
          type: string
          enum: [dark, light]
        dailyNotes:
          type: object
          additionalProperties: false
          properties:
            template:
              type: string
              maxLength: 1024
        editor:
          type: object
          additionalProperties: false
          properties:
            fontSize:
              type: integer
              minimum: 8
              maximum: 32
            lineHeight:
              type: number
              format: double
              minimum: 1
              maximum: 3
            autosaveMs:
              type: integer
              minimum: 250
              maximum: 10000
            showProperties:
              type: boolean
            autoPair:
              type: boolean
            foldGutter:
              type: boolean
            lineNumbers:
              type: boolean
            lineWidth:
              type: integer
              minimum: 400
              maximum: 2000
        accent:
          type: string
          enum: [purple, sky, green, orange]
        readingFont:
          type: string
          enum: [sans, serif]
        server:
          type: object
          additionalProperties: false
          properties:
            port:
              type: integer
              minimum: 1
              maximum: 65535
            dataDir:
              type: string
              minLength: 1
            bind:
              type: string
        mcp:
          type: object
          additionalProperties: false
          properties:
            port:
              type: integer
              minimum: 1
              maximum: 65535
            bind:
              type: string
            auditLog:
              type: boolean
        templates:
          type: object
          additionalProperties: false
          properties:
            folder:
              type: string
              maxLength: 64
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "not_found"
        message:
          type: string
          example: "no note with id 00000000000000000000000001"
...